requirements.txt
altair==4.2.0
backports.zoneinfo==0.2.1
blinker==1.5
bokeh==3.0.3
cachetools==5.2.0
click==8.1.3
contourpy==1.0.6
docopt==0.6.2
et-xmlfile==1.1.0
google-auth==2.15.0
google-auth-oauthlib==0.8.0
gspread==5.7.2
gspread-pandas==3.2.2
heroku==0.1.4
numpy==1.24.1
oauthlib==3.2.2
openpyxl==3.0.10
pandas==1.5.2
Pillow==9.4.0
pipreqs==0.4.11
plotly==5.11.0
protobuf==3.20.3
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.8.0
Pympler==1.0.1
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
requests-oauthlib==1.3.1
rsa==4.9
semver==2.13.0
streamlit==1.16.0
tenacity==8.1.0
toml==0.10.2
toolz==0.12.0
tzdata==2022.7
tzlocal==4.2
validators==0.20.0
watchdog==2.2.1
xyzservices==2022.9.0
yarg==0.1.9
